+++ src/frontend/Phoenix/Home/StatusPanel/StatusPanel.js
@@ -4,9 +4,12 @@
 */
 
 import React from 'react'
+import { Link } from 'react-router-dom'
 import styled from '@emotion/styled'
 import { transparentize } from 'polished'
 
+import superUsersById from '../../../utils/super-users'
+import { useAuth0 } from '../../../../react-auth0-spa'
 import { Colors, Spacing } from '../../../utils/pulseStyles'
 
 // import PushToDevButton from './PushToDevButton'
@@ -52,20 +55,60 @@ const TextLink = styled.a({
   }
 })
 
-const StatusPanel = () => (
-  <Wrapper>
-    <div>
-      <Header>Deploying Changes to Production</Header>
-      <Paragraph>
-        By default, changes automatically appear
-        on <TextLink href="https://dev.pulse-tools.com/" target="_blank">dev.pulse-tools.com</TextLink> after you refresh the Pulse Analytics webapp. For clients to see the changes, click the button below to deploy the changes to production. <TextLink href="https://dedhamgroup.atlassian.net/wiki/spaces/TDG/pages/713129985/Phoenix+User+MGMT#Status-Panel" target="_blank">See guide for help.</TextLink>
-      </Paragraph>
-    </div>
-    {/* <PushToDevButton /> */}
-    <PushToProdButton />
+const StyledButton = styled.button({
+  border: 'none',
+  borderRadius: 4,
+  fontWeight: 700,
+  cursor: 'pointer',
+  padding: `${Spacing.SMALL}`,
+  ':active': {
+    outline: 'none',
+  },
+  ':focus': {
+    outline: 'none',
+  },
+  background: transparentize(0.85, Colors.WHITE),
+  color: Colors.WHITE,
+  ':hover': {
+    background: transparentize(0.7, Colors.WHITE),
+  }
+})
 
-    { process.env.NODE_ENV === 'production' && <OpLog /> }
-  </Wrapper>
-)
+const StatusPanel = () => {
+  const { user } = useAuth0()
+  const isSuperUser = user.sub in superUsersById
+
+  return (
+    <Wrapper>
+      <div>
+        <Header>Deploying Changes to Production</Header>
+        <Paragraph>
+          By default, changes automatically appear
+          on <TextLink href="https://dev.pulse-tools.com/" target="_blank">dev.pulse-tools.com</TextLink> after you refresh the Pulse Analytics webapp. For clients to see the changes, click the button below to deploy the changes to production. <TextLink href="https://dedhamgroup.atlassian.net/wiki/spaces/TDG/pages/713129985/Phoenix+User+MGMT#Status-Panel" target="_blank">See guide for help.</TextLink>
+        </Paragraph>
+      </div>
+      {/* <PushToDevButton /> */}
+      <PushToProdButton />
+
+      { process.env.NODE_ENV === 'production' && <OpLog /> }
+
+      { 
+        process.env.NODE_ENV === 'production' && isSuperUser && (
+          <div style={{ margin: 'auto auto 0px' }}>
+            <div style={{ marginTop: 16 }}>
+              <Link
+                to={`/phoenix/oplog`}
+              >
+                <StyledButton>
+                  Full OpLog
+                </StyledButton>
+              </Link>
+            </div>
+          </div>
+        )
+      }
+    </Wrapper>
+  )
+}
 
 export default StatusPanel